+++ fbink_types.h
@@ -31,18 +31,18 @@
 // List of flags for device or screen-specific quirks...
 typedef struct
 {
-	bool    isPerfectFit;
-	bool    isKindleLegacy;
-	bool    isKindlePearlScreen;
-	bool    isKindleOasis2;
-	bool    isKoboNonMT;
-	bool    isKobo16Landscape;
-	bool    isKoboMk7;
-	int8_t  koboVertOffset;
-	uint8_t koboBootRota;
-	bool    isCervantesNew;
-	uint8_t screenDPI;
-	bool    skipId;
+	bool               isPerfectFit;
+	bool               isKindleLegacy;
+	bool               isKindlePearlScreen;
+	bool               isKindleOasis2;
+	bool               isKoboNonMT;
+	bool               isKobo16Landscape;
+	bool               isKoboMk7;
+	int8_t             koboVertOffset;
+	uint8_t            koboBootRota;
+	bool               isCervantesNew;
+	unsigned short int screenDPI;
+	bool               skipId;
 } FBInkDeviceQuirks;
 
 // An (x, y) coordinates tuple
